Primera iteracion cambio de JS a TS
