+++ package.json
@@ -6,6 +6,10 @@
 		"@testing-library/jest-dom": "^5.16.2",
 		"@testing-library/react": "^12.1.3",
 		"@testing-library/user-event": "^13.5.0",
+		"@types/jest": "^27.4.1",
+		"@types/node": "^17.0.21",
+		"@types/react": "^17.0.39",
+		"@types/react-dom": "^17.0.13",
 		"classnames": "^2.3.1",
 		"lodash.clonedeep": "^4.5.0",
 		"msw": "^0.38.1",
@@ -39,11 +43,15 @@
 		]
 	},
 	"devDependencies": {
+		"@typescript-eslint/eslint-plugin": "^5.13.0",
+		"@typescript-eslint/parser": "^5.13.0",
 		"eslint": "^8.9.0",
 		"eslint-config-airbnb": "^19.0.4",
 		"eslint-plugin-import": "^2.25.4",
 		"eslint-plugin-jsx-a11y": "^6.5.1",
 		"eslint-plugin-react": "^7.28.0",
-		"eslint-plugin-react-hooks": "^4.3.0"
+		"eslint-plugin-react-hooks": "^4.3.0",
+		"ts-loader": "^9.2.7",
+		"typescript": "^4.6.2"
 	}
 }
