+++ src/components/Icons/Icons.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import sprite from '../../assets/img/sprite.svg';
 import PropTypes from 'prop-types';
 
-const Icons = (props) => {
+const Icons = (props: { iconWidth: number; name: string }) => {
 	return (
 		<svg
 			className='search__icon'
