+++ src/components/RectangularBanner/RectangularBanner.tsx
