+++ src/utils/constants/icons.constans.ts
