+++ src/utils/helpers/getDaysBetween.js
@@ -3,10 +3,8 @@ const getDaysBetween = (finalDate) => {
 
 	const oneDay = 1000 * 60 * 60 * 24;
 
-	// Calculating the time difference between two dates
 	const diffInTime = finalDate.getTime() - today.getTime();
 
-	// Calculating the no. of days between two dates
 	const diffInDays = Math.round(diffInTime / oneDay);
 
 	return diffInDays;
