+++ src/utils/helpers/getVotes.ts
@@ -1,4 +1,4 @@
-const getVotes = (positive, negative) => {
+const getVotes = (positive: number, negative: number) => {
 	let iconName = 'thumbs-down';
 	let typeOfButton = 'yellowColor';
 	if (positive > negative) {
