+++ src/utils/helpers/percentage.ts
@@ -1,4 +1,4 @@
-const percentage = (votes, totalVotes) => {
+const percentage = (votes: number, totalVotes: number) => {
 	return Number(Math.round((votes * 100) / totalVotes).toFixed(2));
 };
